requirements.txt
Flask>=3.0.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Werkzeug>=3.0.0
email-validator
Flask-Migrate>=4.0.0
Pillow>=10.0.0